raspberrypi/requirements.txt
sounddevice
scipy
numpy
redis
requests
python-dotenv

